+++ flask_unchained/forms/_compat.py
@@ -1,0 +1,4 @@
+# compatibility with wtforms v3 for wtforms-alchemy
+
+text_type = str
+string_types = (str,)