+++ flask_unchained/forms/fields.py
@@ -1,3 +1,6 @@
 from wtforms.fields import *
-from wtforms.fields.html5 import *
+try:
+    from wtforms.fields.html5 import *
+except ImportError:
+    pass
 from flask_wtf.file import FileField